requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
alibabacloud-credentials==0.3.1
alibabacloud-dysmsapi20170525==2.0.23
alibabacloud-endpoint-util==0.0.3
alibabacloud-gateway-spi==0.0.1
alibabacloud-openapi-util==0.2.1
alibabacloud-tea==0.3.0
alibabacloud-tea-openapi==0.3.6
alibabacloud-tea-util==0.3.8
alibabacloud-tea-xml==0.0.2
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
boto3==1.27.1
botocore==1.30.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==39.0.0
Django==4.1.6
django-cors-headers==3.13.0
django-crontab==0.7.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
et-xmlfile==1.1.0
frozenlist==1.3.3
idna==3.4
jmespath==1.0.1
multidict==6.0.4
mysqlclient==2.2.4
openpyxl==3.1.0
pycparser==2.21
pycryptodome==3.20.0
PyJWT==2.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2022.7.1
requests==2.28.2
s3transfer==0.6.2
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14
yarl==1.8.2
